+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Change types: remove `MigrationHistory`, add `MigrationRecordInfo` & `MigrationSourceInfo` [#32](https://github.com/Synor/core/issues/32)
 - Combine `Migrator.history` and `Migrator.pending` into `Migrator.info` [#32](https://github.com/Synor/core/issues/32)
 - Add support for Out of Order migration run [#32](https://github.com/Synor/core/issues/32)
+- Add `extension` property in `MigrationInfo`
 - ...
 
 ## [0.7.0] - 2020-01-14